feat: made claim property view reusable
