+++ src/app/(light)/dashboard/components/claim-view/claim-view.tsx
@@ -1,0 +1,38 @@
+import { pda as pdaLocale } from '@/locale/en/pda';
+import { CredentialData } from '@/services/protocol/types';
+import getClaimType, { ClaimField } from '@/utils/get-claim-type';
+
+import { CurrencyView } from './currency-view';
+import { ImageView } from './image-view';
+import { LinkView } from './link-view';
+import { ListView } from './list-view';
+
+export default function ClaimView(fieldData: CredentialData) {
+  const type = getClaimType({
+    type: fieldData.type,
+    contentMediaType: fieldData.metadata?.contentMediaType,
+    currency: fieldData.metadata?.currency,
+    format: fieldData.metadata?.format,
+  });
+
+  if (!fieldData.value || fieldData.value === '')
+    return <span>{pdaLocale.unfilled}</span>;
+  switch (type) {
+    case ClaimField.Image:
+      return <ImageView src={fieldData?.value} alt={fieldData?.label} />;
+    case ClaimField.Array:
+      return <ListView value={fieldData?.value} />;
+    case ClaimField.Link:
+      return <LinkView href={fieldData?.value} />;
+    case ClaimField.Currency:
+      return (
+        <CurrencyView
+          currency={fieldData?.metadata?.currency}
+          value={parseFloat(fieldData?.value) as number}
+        />
+      );
+    default: {
+      return <span>{fieldData.value}</span>;
+    }
+  }
+}
