+++ src/app/(light)/dashboard/components/claim-view/currency-view.tsx
