+++ src/app/(light)/dashboard/components/claim-view/image-view.tsx
